Fixed a couple of things. Started trying to make some bindings for the library so that we can use the library from the test program. Not done yet, I'm a little dissatisfied with how the bindings are turning out. I'll have to think about it some more and do some research into how OpenGL and such things do it. This commit is bad, haven't compiled, doesn't work, all that. Don't use this one.

+++ src/lib/lexer.h
@@ -18,13 +18,13 @@ class lexer_t {
 	size_t current_stream_position = 0;
 
 	dive_token_t current_token {
-		-1,
+		(size_t)-1,
 		current_stream_position,
 		0
 	};
 
 	dive_token_t last_token {
-		-1,
+		(size_t)-1,
 		0,
 		0
 	};
